split components into separate packages
