+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "lerna": "2.9.0",
+  "lerna": "2.11.0",
   "packages": [
     "packages/*"
   ],
