+++ packages/react-slate-components/.eslintignore
@@ -1,0 +1,3 @@
+**/node_modules/**
+**/flow-typed/**
+**/build/**
