+++ packages/react-slate-components/README.md
@@ -1,4 +1,4 @@
-# @react-slate/utils
+# @react-slate/components
 
 [![Version][version-badge]][package]
 [![Build Status][build-badge]][build]
@@ -19,7 +19,7 @@ Prototype of [Haul CLI](https://github.com/callstack/haul). You can see the [sou
 ### Installation
 
 ```bash
-yarn add @react-slate/core @react-slate/utils
+yarn add @react-slate/core @react-slate/components
 ```
 
 ### Documentation
@@ -30,10 +30,10 @@ Documentation can be found here: http://react-slate.surge.sh/
 
 [build-badge]: https://img.shields.io/circleci/project/github/zamotany/react-slate/master.svg?style=flat-square
 [build]: https://circleci.com/gh/zamotany/react-slate
-[license-badge]: https://img.shields.io/npm/l/@react-slate/utils.svg?style=flat-square
+[license-badge]: https://img.shields.io/npm/l/@react-slate/components.svg?style=flat-square
 [license]: https://opensource.org/licenses/MIT
-[version-badge]: https://img.shields.io/npm/v/@react-slate/utils.svg?style=flat-square
-[package]: https://www.npmjs.com/package/@react-slate/utils
+[version-badge]: https://img.shields.io/npm/v/@react-slate/components.svg?style=flat-square
+[package]: https://www.npmjs.com/package/@react-slate/components
 [prs-welcome-badge]: https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square
 [prs-welcome]: http://makeapullrequest.com
 [coc-badge]: https://img.shields.io/badge/code%20of-conduct-ff69b4.svg?style=flat-square
