+++ packages/react-slate-components/src/index.js
@@ -1,0 +1,4 @@
+/* @flow */
+
+export { default as ProgressBar } from './components/ProgressBar';
+export { default as Spinner } from './components/Spinner';
