--- packages/react-slate-core/src/components/index.js
@@ -1,6 +1,0 @@
-export { default as View } from './View';
-// For compatibility
-export { default as Text } from './View';
-export { default as KeyPress } from './KeyPress';
-export { default as Spinner } from './Spinner';
-export { default as ProgressBar } from './ProgressBar';
