+++ packages/react-slate-core/src/index.js
@@ -2,7 +2,6 @@
 
 import onExit from './utils/onExit';
 
-export * from './components';
 export { default as colors } from './constants/colors';
 export {
   default as renderToTerminal,
@@ -10,6 +9,9 @@ export {
 } from './public/renderToTerminal';
 export { default as renderToString } from './public/renderToString';
 export { default as renderToTarget } from './public/render';
+export { default as View } from './public/View';
+// For compatibility
+export { default as Text } from './public/View';
 
 export const Internal = {
   onExit,
