+++ packages/react-slate-core/src/public/View.js
