+++ packages/react-slate-interactive/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@react-slate/core",
-  "version": "0.5.1",
-  "description": "Render React CLI apps to TTY terminals",
+  "name": "@react-slate/interactive",
+  "version": "0.2.0",
+  "description": "Components to interact with TTY in react-slate apps",
   "license": "MIT",
   "repository": "https://github.com/zamotany/react-slate",
   "private": false,
@@ -26,24 +26,14 @@
     "prepublish": "yarn build"
   },
   "peerDependencies": {
-    "react": "16.x"
+    "react": "16.x",
+    "@react-slate/core": "*"
   },
   "dependencies": {
-    "ansi-parser": "^3.2.8",
-    "ansi-styles": "^3.2.0",
-    "astral-regex": "^1.0.0",
-    "chalk": "^2.1.0",
-    "cli-spinners": "^1.1.0",
-    "fast-memoize": "^2.2.8",
-    "fbjs": "^0.8.16",
-    "is-fullwidth-code-point": "^2.0.0",
-    "mkdirp": "^0.5.1",
-    "prop-types": "^15.6.0",
-    "react-reconciler": "0.3.0-beta.1",
-    "strip-ansi": "^4.0.0"
   },
   "devDependencies": {
-    "react": "^16.3.2"
+    "react": "^16.3.2",
+    "@react-slate/core": "*"
   },
   "babel": {
     "presets": [
