+++ packages/react-slate-interactive/src/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { default as KeyPress } from './components/KeyPress';
