+++ packages/react-slate-playground/src/App.js
@@ -1,5 +1,7 @@
 import React from 'react';
-import { View, Spinner, ProgressBar, KeyPress } from '@react-slate/core';
+import { View } from '@react-slate/core';
+import { Spinner, ProgressBar } from '@react-slate/components';
+import { KeyPress } from '@react-slate/interactive';
 
 const components = ['Spinner', 'ProgressBar'];
